Create arithmetics.r
var1 is logical. Convert it to character and assign it to the variable var1_char.
Next, see whether var1_char actually is a character by using the is.character() function on it.
var2 is numeric. Convert it logical and assign it to the variable var2_log.
Inspect the class of var2_log using class().
Finally, var3 is of type character. Convert it to numeric and assign the result to var3_num. Was it successful?