Create Apples-and-Oranges.r
Common sense tells you not to add apples and oranges. The my_apples and my_oranges variables both contained a number in the previous exercise. The + operator works with numeric variables in R. If you really tried to add "apples" and "oranges", and assigned a text value to the variable my_oranges but not to my_apples (see the editor), you would be trying to assign the addition of a numeric and a character variable to the variable my_fruit. This is not possible.